sem: rename to sema
To prevent name colisions with POSIX semaphores
